back-end/requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-appconf==1.0.6
django-cors-headers==4.3.1
django-filter==23.5
django-staticfiles==1.2.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.2
pillow==10.2.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
sqlparse==0.4.4
stripe==8.3.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
